+++ src/evaluator.h
@@ -3,7 +3,7 @@
 #include "environment.h"
 #include "parser.h"
 
-int eval(ASTNode *node, Environment *env);
+long eval(ASTNode *node, Environment *env);
